a few more -Wconversion issues
git-svn-id: https://svn.r-project.org/R/trunk@59202 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/RNG.c
@@ -168,7 +168,7 @@ double unif_rand(void)
 	if (p2 < 0.0) p2 += m2;
 	II(3) = II(4); II(4) = II(5); II(5) = (int) p2;
 
-	return ((p1 > p2) ? (p1 - p2) : (p1 - p2 + m1)) * normc;
+	return (double)((p1 > p2) ? (p1 - p2) : (p1 - p2 + m1)) * normc;
     }
     default:
 	error(_("unif_rand: unimplemented RNG kind %d"), RNG_kind);
